clean up code and files
