--- examples/buffer-example/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "buffer_example",
-  "version": "0.0.1",
-  "private": true,
-  "gypfile": true,
-  "scripts": {
-    "start": "node index.js"
-  },
-  "dependencies": {
-    "nan": "*",
-    "stream-buffers": "^3.0.1"
-  }
-}
